change: fix form loader and test on base project
