+++ lib/components/Loaders/FormLoader/FormLoader.tsx
@@ -1,26 +1,13 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import React from 'react'
-
-interface RowLoaderProps {
-  cols: number // Number of columns for the row
-}
-
-export const RowLoader: React.FC<RowLoaderProps> = ({ cols }) => {
-  return (
-    <div className="flex space-x-2 mb-2">
-      {/* Generate columns based on cols for the row */}
-      {Array.from({ length: cols }, (_, colIndex) => (
-        <span key={colIndex} className="animate-pulse block h-8 rounded-md w-52 bg-cu-black-100"></span>
-      ))}
-    </div>
-  )
-}
+import { RowLoader } from './RowLoader' // Import RowLoader component
 
 interface FormLoaderProps {
-  children: any
+  children: React.ReactNode
 }
 
-export const FormLoader: React.FC<FormLoaderProps> = ({ children }) => {
+// Define FormLoader component
+export const FormLoader: React.FC<FormLoaderProps> & { RowLoader: React.FC<any> } = ({ children }) => {
+  // Return FormLoader with children and RowLoader as property
   return (
     <>
       {React.Children.map(children, (child, index) => (
@@ -30,3 +17,6 @@ export const FormLoader: React.FC<FormLoaderProps> = ({ children }) => {
     </>
   )
 }
+
+// Assign RowLoader to FormLoader
+FormLoader.RowLoader = RowLoader
